Added some sprite bounds for default bounds.

+++ Code/Engine/Renderer/2D/SpriteGameRenderer.cpp
@@ -115,7 +115,7 @@ void SpriteLayer::CleanUpDeadParticleSystems()
 //-----------------------------------------------------------------------------------
 SpriteGameRenderer::SpriteGameRenderer(const RGBA& clearColor, unsigned int widthInPixels, unsigned int heightInPixels, unsigned int importSize, float virtualSize) : m_clearColor(clearColor)
     , m_importSize(importSize) //Artist's asset size. How big do you make the assets? 240p, 1080p, etc... (144p for gameboy zelda)
-    , m_aspectRatio(widthInPixels/heightInPixels)
+    , m_aspectRatio((float)widthInPixels/(float)heightInPixels)
     , m_defaultRenderState(RenderState::DepthTestingMode::OFF, RenderState::FaceCullingMode::CULL_BACK_FACES, RenderState::BlendMode::ALPHA_BLEND)
     , m_cameraPosition(Vector2::ZERO)
     , m_worldBounds(-Vector2::MAX, Vector2::MAX)
